text2sql_fastapi/requirements.txt
agno
aiohappyeyeballs
aiohttp
aiosignal
aiosmtplib
alembic
amqp
annotated-types
anyio
asgiref
asyncmy
asyncpg
attrs
bcrypt
billiard
blinker
cachetools
celery
certifi
cffi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
cryptography
dataclasses-json
DateTime
distro
dnspython
docstring_parser
email_validator
elevenlabs
fastapi
fastapi-mail
frozenlist
gitdb
GitPython
google-auth
google-genai
greenlet
groq
h11
httpcore
httpx
httpx-sse
idna
Jinja2
jsonpatch
jsonpointer
kombu
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
langsmith
lxml
Mako
markdown-it-py
MarkupSafe
marshmallow
mdurl
multidict
mypy_extensions
numpy
orjson
packaging
passlib
prompt_toolkit
propcache
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
PyJWT
PyMySQL
python-dateutil
python-docx
python-dotenv
python-multipart
pytz
PyYAML
redis
requests
requests-toolbelt
rich
rsa
setuptools
shellingham
six
smmap
sniffio
SQLAlchemy
sqlmodel
starlette
tenacity
tomli
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
upstash-redis
urllib3
uvicorn
vine
wcwidth
websockets
yarl
zope.interface
zstandard
PyMuPDF
Pillow
llama-cloud-services
fastapi-mail

